+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+
+## 0.4.8
+Fixed an issue where BLE devices would disappear from the device list on iOS, once connected.
+
+## 0.4.7
+Added null_checks for latest flutter sdk (kotlin)
+fixed NRPN / added RPN
+
+## 0.4.6
+Fixed an issue showing the OS bluetooth permission alerat before
+refresh button was pressed.
+
+## 0.4.5
+Added Pitch bend message type
+More controls in example app
+
+## 0.4.4
+Fixed issue with single messages, credit to https://github.com/anzbert
+
+## 0.4.3
+added android:exported="true" to manifest
+
 ## 0.4.2
 Fixed BLE congestion issues on iOS